chore: switch to abstract sqlite class
